requirements.txt
python-telegram-bot==20.6
supabase==1.0.3
aiohttp==3.8.5